requirements.txt
selenium==3.3.1
jsonschema==2.6.0
Appium-Python-Client==0.14
pytest==3.1.3
pytest-html==1.15.1
pytest-json==0.4.0

